wrap `detectChanges` in a setTimeout call
There is a [bug](https://github.com/angular/angular/issues/38611) in the Angular framework that impact some applications that use this library with Ivy.

Due to that [bug](https://github.com/angular/angular/issues/38611#issuecomment-683277944), a runtime error is thrown when calling `detectChanges` inside the `transform` method of the `LocalizeRouterPipe`.